[PATCH] OMAP: DSP pm fix
This is a fix for pm code with regard to the DSP.

In the current code we have a problem that DSP won't 
re-enabled after suspend if DSP is just turned on
(with idle request) from McBSP  driver but not used
by DSP Gateway.

I moved DSP related code from pm.c to dsp_common.c.
With this patch DSP will be resumed to the original state 
after suspend, whichever state (reset / idling for
McBSP / running with DSP Gateway) DSP was in.

Signed-off-by: Toshihiro Kobayashi <Toshihiro.Kobayashi@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>