[PATCH] ARM: OMAP: make Mistral support explicit
This starts packaging support for the Mistral board and its devices
separately from OSK framebuffer support, and supports the "wakeup"
button (SW1) on that board.  This involved some mux fixes, both for
the MicroWire stuff (!) and for MPUIO2.

There's still a lot more to be done ... currently the board-specific
setup for the LCD, keypad, touchscreen, and so forth is scattered
throughout several drivers, rather than gathered in one place where
it's more naturally updated to match board revisions (or replaced
for folk using different UI boards).

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Tony Lindgren <tony@atomide.com>