[PATCH] ARM: OMAP: Change return from address to content in dma.c
The following patch changes the return value from functions :

- omap_get_dma_src_pos
- omap_get_dma_dst_pos

in file dma.c.

The original version of the functions returns the address of the
registers that has the related information. The new version returns
the content of the registers: the source and destination addresses of
the current transfer for the given channel.