[PATCH] ARM: OMAP: MMC multi-block fix
Cope with a glitch in the mmc block layer, needed whenever multiblock
writes are disabled.  In that case, the mmc layer passes scatterlists
that ask for (a) multiple segments, (b) which may be multiple blocks.
Both must be ignored with CMD24, "single block write".

Start dma operations only after writing the command.  This is a better
match for the specs; it matches what PIO does, and could address the
lockups seen on some H3 boards.

Minor debug cleanups of the MMC driver.