[PATCH] ARM: OMAP: USB peripheral DMA update'
The change to make DMA work two bytes at a time omitted an important
tweak that affects the file_storage gadget:  it needs to recognize when
the host writes an odd number of bytes.  (The change was made to get
past a controller lockup that seemed to be caused by byte-at-a-time DMA
using too much bandwidth on one of the I/O busses during network stress
testing.  The network layer ignores such extra bytes.)

This patch resolves that issue by checking the relevant bit and adjusting
the rx byte count, so that for example a legal 13 byte request doesn't
morph into an illegal 14 byte one any more.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Tony Lindgren <tony@atomide.com>