[PATCH] ARM: OMAP: mux: simplify MUX_CFG_730 and add some usb client pins
As I continue to try to sort out USB gadget support on the 730, I find
that I need to add some pin definitions.

Since the 730 lacks PU_PD registers, and the pullup bit is always in
the config register on the 730, I simplified the MUX_CFG_730() macro
a little bit (to avoid typing a bunch of 0s and NAs on *every* line).