[PATCH] ARM: OMAP: Remove support for board Netstar
Remove support for board Netstar. I hope it helped to sort out various
needs to drivers and now when nearly everything was converted to
platform device there is no more need to bother you with updates to
board file :-). I'll maintain it as separate patch (using PTXdist).
Thanks for all the fish :-)