[PATCH] ARM: OMAP: STI/XTI support
This implements a simple subsystem for the OMAP serial tracing interface,
as found in omap16xx and omap24xx.

There's a few things implemented in this patch:

        - STI console support
        - STI netlink support
        - STI RX FIFO support

In addition to this, we also provide a general tracing framework that can
be used by both the kernel and userspace (and imask control for allowing
DSP-side STI manipulation).

Since the IRQ is multiplexed, drivers register for the bits they're
interested in (the RX FIFO is an example of this).

The netlink interface exists for streaming data bi-directionally. This is
currently done using NETLINK_USERSOCK, which ends up being somewhat
limiting.

All of these things have been tested on both OMAP1 and OMAP2, and it
takes in to account some of the protocol differences.