[PATCH] ARM: OMAP: USB gadget driver
While doing development for an OMAP2420 board I noticed that the
"workaround for seemingly-lost IRQs for RX ACKs" in omap_udc.c needs a
little tweak in order to prevent NFS root over usbnet from hanging.
According to the documentation the status flag register should only be
accessed when endpoint select bit is set in the EP_NUM register. See
e.g. SPRU761A, end of page 78 (about STAT_FLG register):

"This register cannot be read if EP_NUM.EP_SEL bit is not asserted for the
endpoint."

Similar statement seems to be in the specs for other OMAPs too.

As a change for the workaround seem to have an effect it seems that
there is indeed an underlying "IRQs get lost" problem?

In principle reading the status flag in proc_ep_show() would need
similar change, but I'm not sure how the EP selection would interact
with interrupt processing. Actually I'm not sure how this change
interacts with interrupt processing either but at least it cured the
hangs for the 2420.

It seems that usbnet with Nokia 770 is also a bit unreliable but this
change does not seem to have an effect. Does someone have an OMAP
based system where usbnet is reliable?