ARM: OMAP: SoSSI: workaround for HW timing bug
Certain HW timings can render the SoSSI block dead. This can happen when
the SoSSI TW1 value is less then 3 and we are trying to read the SoSSI
FIFO. At this point SoSSI goes haywire and return 0 for all its
registers. This patch will set a minimum limit of 3 on TW1.

Signed-off-by: Imre Deak <imre.deak@nokia.com>
Signed-off-by: Juha Yrjola <juha.yrjola@nokia.com>