[PATCH] ARM: OMAP: Alsa modularisations and support for tsc2101 3/3 (round 2)
This 3/3 patch adds the tsc2101 alsa driver support for the omap-h2.

Just like the aic23 alsa driver, the codec specific functionality is
separated from the common alsa driver functionality that is in the omap-alsa.c.
tsc2101 mixer supports currently the playback selection between
loudspeaker and headset/handset.
Addon of others input and targets should be possible to do afterwards.

In addition it should be possible to add later the support for the
dynamic clocks on top of this series of patches, as discussed
in the omap-linux mailing list.

Signed-off-by: Mika Laitio <lamikr@cc.jyu.fi>
Signed-off-by: Daniel Petrini <d.pensator@gmail.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>