[PATCH] ARM: OMAP: Add support for Amstrad Delta (E3) Videophone
This patch adds support for the Amstrad Delta (E3) Videophone. This is
an OMAP5910 based device, which you can find more details about at:

http://www.amstrad.com/products/emailers/e3.html
http://www.earth.li/~noodles/hardware-e3.html

Patch is against latest git://source.mvista.com/git/linux-omap-2.6 and
adds the bare minimum of support for the device; I have a handful of
other patches to sort out the NAND, LCD, keymap and USB but they all
depend on this one.

Signed-Off-By: Jonathan McDowell <noodles@earth.li>
Signed-Off-By: Tony Lindgren <tony@atomide.com>