[PATCH] ARM: OMAP: dspgateway sem2mutex conversion
This has been lagging behind a bit, primarily due to the fact that the
OMAP tree has gotten a bit out of sync with the DSP gateway updates
(we'll leave that for another patch).

This is a quick and simple sem2mutex conversion, after this dspgateway is
essentially down to a single rwsem for the exmap cases, and that will
remain as it is.

Against current git.

Signed-off-by: Paul Mundt <paul.mundt@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>