[PATCH] omap_uwire: use the per transfer wordsize and speed overrides
Each transfer in an SPI message can have its wordsize and speed override.
Add a uwire_setup_transfer to handle this.

Allocate a per SPI device controller state struct, where we cache the
current bits_per_word and the functional clock divider. The latter cannot
be set for each CS in HW, so we have to reset it whenever a new device is
selected.

Signed-off-by: Imre Deak <imre.deak@nokia.com>