[PATCH] ARM: OMAP: Fix checking for multiple boot tags in board config
Currently if we look for more than one instance of a boot tag and it's
defined in the board config rather than passed by the boot loader we can
get stuck in a loop where the same tag is processed many times. This can
be seen with the gpio-switch driver, for example. The patch below fixes
this by respecting the skip parameter to get_config when we're scanning
the board config structure.

Signed-Off-By: Jonathan McDowell <noodles@earth.li>