[PATCH] ARM: OMAP: add-RGB444-support-to-omapfb.patch
This is an updated version of my patch that adds support for RGB444 mode
for the internal LCD controller which now applies against linux-2.6.16
with the patch-2.6.16-omap2 applied.

Jonathan McDowell posted a slightly modified version of my original patch
to this list previously:

http://Linux.omap.com/pipermail/linux-omap-open-source/2006-April/006839.html

This patch is used by the Amstrad Delta (E3).

Signed-off-by: Mark Underwood <basicmark@yahoo.com>
Acked-by: Jonathan McDowell <noodles@earth.li>
Signed-off-by: Imre Deak <imre.deak@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>