RTC: Add rtc-omap
This creates a new RTC-framework driver for the RTC/calendar module
found in various OMAP chips, giving a more correct/standard replacement
for the older drivers/char/omap-rtc.c driver.  Differences include:

 - much smaller/simpler, because it reuses shared infrastructure
 - the RTC name will normally be "rtc0" not "rtc"
 - the /dev node has a different major and minor numbers
 - RTC_ALM_SET handled as on PCs (alarm within 24 hours)
 - RTC_WKALM_SET handled as on PCs (alarm within this century)
 - epoch not changeable (why bother)
 - rtc alarm may optionally be a system wakeup event (board-specific)

If you use udev and statically link this, some init script can just create
a symbolic link (rtc -> rtc0) so tools like "hwclock" will work as usual.
Or, updates to hwclock (and busybox) are available which teach it to use
the /dev/rtc0 path as a backup, and to support "hwclock --file=/dev/rtc0".

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>