SPI: TSC2102 OMAP audio driver
ALSA driver for OMAP boards with a TSC2102 chip. It is based on the
TSC2101 and AIC23 drivers. The chip has only one audio playback output
and no inputs. It can play stereo sound with 16-bit samples (only) and
supports different sampling rates between 8 and 48 kHz. The patch also
includes an ALSA mixer. The mixer controlls two volume levels, for
left and right channels, each with 128 steps. It can also toggle the
analog de-emphasis filter and bass boost filter which are incorporated
in the chip.

In addition to that the chip has an equaliser filter, but the mixer
doesn't use it because the coefficients of the filter are difficult to
map to other values used in normal equalisers. Maybe they should have
constant values per board.

Signed-off-by: Andrzej Zaborowski <balrog@zabor.org>