ARM: OMAP: Set serial 8250 ECB for TX fifo
Code to fix serial TX fifo by Richard Woodruff.

The extended capabilities bit is needed to write fcr, or else some
settings don't get set.  That is what the OMAP-UART data sheet says.
The write water mark bits 4:5 are not accessible unless ECB is set.

The data sheet says it is TI16C750 compatible but it seems to share more
with what the code calls a TI16C752.  The 750 has a DLAB to set 64bit
fifo and I don't believe that is the case for ours.

A minor note is we are still only using 32 bytes of the 64 byte TX fifo
for transmits in the current code.  I'm not sure if there is an
application that this matters to.

Note that the current code id's OMAP serial ports as a 16654 which isn't
quite correct.

Also note that to use the actual TX watermark interrupt set at 32bytes
required too ugly of a hack to use (not use OMAP_SCR=0).