ARM: OMAP: Export reserved APIs in GPMC header
The attached patch is to export two APIs (gpmc_cs_set_reserved and
gpmc_cs_reserved) in gpmc.h.

This allows other libraries and drivers to access these APIs with out
doing a gpmc_request.

Reason: If I have already configured a GPMC CS in boot loader I might
not be interested in doing the same again in Kernel. However I might
want to reserve this particular CS so that it doesn't show as free when
others do a gpmc_request.

Signed-off-by: Syed Mohammed Khasim <x0khasim@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>