ARM: OMAP: hrtimers/dyntick support for OMAP2 GP-timers
This patch converts the GP-timer layer to use the
clocksource/clockevent interface provided by the new GENERIC_TIME and
GENERIC_CLOCKEVENT layers.

Doing so enables use of the new hrtimers and dynticks features on
OMAP2.

A quick test of hrtimers using Thomas Gleixner's cyclictest on a
2430SDP shows min/avg/max userspace latencies of 38/50/850 usec.

A quick test of dyntick using /proc/timer_stats on an idle system
shows an avg. of ~5 events/sec over a 10 second period.

Signed-off-by: Kevin Hilman <khilman@mvista.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>