OMAP Backlight driver
in the actual repository the video backlight API is already updated to the new
one in kernel 2.6.22-rc1 (include/linux/backlight.h).
Some drivers are updated,too.
However, the backlight driver for Omap (drivers/video/backlight/omap_bl.c) ist
not yet updated to the new API and for this reason cannot be compiled.

I changed the file according to the new interface

Signed-off-by: Stefano Babic <stefano.babic@babic.homelinux.org>
Signed-off-by: Tony Lindgren <tony@atomide.com>