tsc2102-alsa build fixes
Various fixes needed to build the tsc2102 audio.  Several fixes
relate specifically to building it as a module.

Note that this doesn't address the need for omap-alsa-dma.c
to go into a separate module so that e.g. tsc2101 and tsc2102
modules can both be built, or the assumption that the audio
is going out over McBSP not EAC.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>