ARM: OMAP: fix clock table walking for OMAP2 variants
The various places where the clock rate_table is walked are not all
checking whether the clock actually exists on the OMAP2 variant.

Among other things, this results in broken rounding for the OMAP2430
since the rounding code can potentially find matching speeds that are
marked as RATE_IN_242X.

It also results in confusing CPUfreq about what the current speed is.

Signed-off-by: Kevin Hilman <khilman@mvista.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>