omap2 clock: clean up dss2_fck clock flags
dss2_fck is not a fixed-rate clock; its rate is dependent on its parent's
rate.  So, drop the .rate assignment and drop the RATE_FIXED clock flag.
It also is a leaf clock - no clocks have it as a parent - so it can't
propagate its rate anywhere.  So, drop the RATE_PROPAGATES and fix the
recalc function pointer accordingly.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>