omap2 clock: remove superfluous omap2_propagate_rate()
omap2_propagate_rate() seems completely superfluous.  It handles two
cases.  The first case is for non-fixed rate clocks.  It assigns the
parent's rate to the clock, and continues propagating the rate
downwards.  This can be handled by adding a RATE_PROPAGATES clock flag
to the clock, if it wasn't there already, and by using
followparent_recalc() directly as the recalc function.  The second
case is for fixed rate clocks.  Here all we need to do is to call
propagate_rate() directly, so, use that for the recalc function.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>