Remove unused and broken core_l3_ck clk_get from board-3430sdp.c
board-3430sdp.c tries to clk_get core_l3_ck.  This clock doesn't exist
any longer in the current clock34xx.h, so it fails.  The code then tries
to fetch the clock's rate, which is also not currently used.  This is
dead code, so remove it.

Tested on 3430SDP ES1.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>