Fix MUSB DMA for inbound ISO endpoints
The DMA receive path in the MUSB driver does not update the fields in
the URB for ISO end points (the PIO path does, however). This patch
updates the fields in the URB for the DMA path. Drivers using ISO endpoints
will not see any valid data without those fields since the USB core initializes
them to be zero/error.

Tested with on the 2430SDP with a USB bluetooth dongle and SCO audio
recording and playback.

This implements the ISO URB handling for the DMA rx case by providing the
appropriate length and status in the struct usb_iso_packet_descriptor field of
the URB. Previously, this was only done in the PIO case.

Signed-off-by: Hunyue Yau <hyau@mvista.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>