T2: Remove redundant clear and restore of the COR flag
The T2 Clear-On-Read (COR) feature works by clearing all pending
irqs when the ISR register is read. It should be needless to say
that this feature isn't compatible with shared interrupts, unless
the value of the ISR is read only once for each interrupt.

It is therefor safe to assume that COR is not used - because if a
driver uses the feature, that driver will clear the interrupts
belonging to other drivers.

This patch removes the redundant clearing and restoring of the COR
bit in REG_PWR_SIH_CTRL.

Signed-off-by: Klaus Pedersen <klaus.k.pedersen@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>