Wait for OMAP3430 clocks to enable before returning from clk_enable()
This patch fixes some boot-time instability problems with OMAP3430. The
GPTIMER modules can take 200 to 300 microseconds to become available after
their clocks are enabled, and without this patch, the clock framework
won't wait for them before returning from clk_enable().  If the system
attempts to access a GPTIMER register too quickly after the clk_enable(),
the system will halt with something similar to the following in the
dmtimer code:

Unhandled fault: external abort on non-linefetch (0x1028) at 0xd9032010

This is perhaps not the ideal way to solve this problem, but then again,
omap2_clk_wait_ready() is perhaps not the ideal way to wait for module
readiness.  This function is due to be replaced, but in the interim, this
patch fixes the issue.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>