tsc2301 - add coordinate average filter
To get good touchscreen results filtering is essential.
The filter does 2 things - it removes "bad" measurements,
and it uses average to smooth the coordinates.

Use the HW-filter in the tsc2301 to remove high frequency
noise (from backlight). Using 8 sample average is very
efficient.
Another source of noise is the screen, but that generates
low freq noise and thus require a SW filter. Using 4
coordinate history average is efficient.

Signed-off-by: Klaus Pedersen <klaus.k.pedersen@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>