tsc2301 - correct the handling of the keyb_int interrupt
This patch fixes the "geiger" sound that was generated when
touching the touchscreen and pressing a key at the same time.

The error was the use of the keyb_int interrupt line as a
shortcut to see if any keys were pressed, that doesn't work.

With the tsc2301 it is essential that the keypad status is read
soon after the interrupt arrive - otherwise you will occasionally
read the keyboard status while the keys are still being de-bounced
(nothing).

Signed-off-by: Klaus Pedersen <klaus.k.pedersen@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>