OMAP: MMU: Fix inconsistency to unregister device MMU
There was a problem that kernel counldn't find struct device for
device MMU when unregistering it for class. This is because device
MMUs are the pseudo devices and they were not created when registering
into MMU class. This patch will create struct device when registering
device MMU into class and will fix the problem that "dspctl cleanup"
encountered kernel Oops.

Signed-off-by: Hiroshi DOYU <Hiroshi.DOYU@nokia.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>