ARM: OMAP: HSMMC: enable use as a module
When building as a module, the board support code (which is compiled
in) cannot directly call the driver code (which may be in a module.)

This patch the separates the card-detect IRQ usage into board-specific
code and driver code, and adds a couple slot-specific items to the MMC
platform data.

Tested on 3430SDP ES2.

Signed-off-by: Kevin Hilman <khilman@mvista.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>