OMAP2/3 clock: convert mask_to_shift() to __ffs()
In OMAP2/3 clock code, we've used mask_to_shift() to convert bitmasks
into shift values, via "ffs(mask) - 1".  It turns out that there is
already a Linux idiom for this in asm/bitops.h: __ffs().  (Not to be
confused with ffs(), of course.  You wouldn't do that, would you?)
When in Rome, do as the Romans.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>