OMAP3 PM: enable hardware SAR for USBHOST, USBTLL modules
Enable hardware save-and-restore for the CORE and USBHOST powerdomains
during PM layer initialization.  On OMAP3, this only affects the
USBTLL and USBHOST modules.

There is probably a sleep and wakeup latency penalty with these
enabled.  No one seems to have quantified it yet.  If the added
latency is measurable, an alternate approach would be to only enable
hardware save-and-restore if there are USB devices attached.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>